Refine DataEngine organization
